Begin splitting out components
