+++ gemp-lotr/gemp-lotr-async/src/main/web/index.html
@@ -38,7 +38,7 @@
 	<div id="root"></div>
 
 	<!-- vite bundles -->
-    <script type="module" crossorigin src="dist/assets/main-Dw1HkOIO.js"></script>
+    <script type="module" crossorigin src="dist/assets/main-BJM1IVBx.js"></script>
     <link rel="stylesheet" crossorigin href="dist/assets/main-BVyQIhOb.css">
 
 </body>
